requirements.txt
beautifulsoup4==4.12.2
pandas==2.0.3
selenium==4.11.2
pyaml==6.0.1
requests==2.31.0